if indentation is less than zero set it to zero.
2002-05-15  Larry Ewing  <lewing@ximian.com>

	* htmlclueflow.c (html_clueflow_set_indentation): if indentation
	is less than zero set it to zero.

	* htmlengine-edit-clueflowstyle.c (get_props_and_set): move the
	indentation changes before the style changes and explicitly set
	indentation for items.  As long as indentation is changed before
	the item type we can correctly set the toplevel value of the
	indentation array.

	* htmlclueflow.h: change set_indentation prototype.

	* gtkhtml.c (paragraph_style_to_clueflow_style): make the default
	style be blockquote, and make item dotted be unordered.

svn path=/branches/larry-multicite-branch/; revision=5034

+++ src/htmlclueflow.h
@@ -91,7 +91,7 @@ void               html_clueflow_modify_indentation_by_delta  (HTMLClueFlow     
 							       gint                indentation);
 void               html_clueflow_set_indentation              (HTMLClueFlow       *flow,
 							       HTMLEngine         *engine,
-							       guint8              indentation);
+							       gint              indentation);
 guint8             html_clueflow_get_indentation              (HTMLClueFlow       *flow);
 void               html_clueflow_set_properties               (HTMLClueFlow       *flow,
 							       HTMLEngine         *engine,
